Project: Notes Application Works Cloud Storage Provided by Firebase
Introduction
* Read more about Cloud Storage for Firebase
Getting Started
* Add Firebase to your Android Project. The package name you'll use is package naveen.agnosbeta.notepad.
* Run the Android application on your Android device or emulator.
* Upload Any type of files using this app.
* Same data can be shared by number of users at the same time in synchronised manner.
* Also provide other features like upload images or videos via camera.
* User can also download the files and delete the uploaded files from the cloud directly using the same application on android.

Support
https://firebase.google.com/support/
License
Copyright 2018 Google, Inc.
Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.